- move DFA.equals method into a function in FSMEquality.kt for better readability - remove unnecessary backing fields in State
